OpenBSD FFM port: move native-free code to oshi-common (#3332)
Move pure-Java OpenBSD classes (no JNA/native dependencies) from
oshi-core to oshi-common so they can be shared by both oshi-core (JNA)
and oshi-core-ffm (FFM). Mirrors the FreeBSD port structure from #3316.

Moved to oshi-common:
- FstatUtil, OpenBsdOSThread, OpenBsdOSProcess (abstract base),
  OpenBsdOSFileStore, OpenBsdFileSystem (abstract base),
  OpenBsdInternetProtocolStats, OpenBsdNetworkParams,
  OpenBsdFirmware, OpenBsdGraphicsCard, OpenBsdSoundCard,
  OpenBsdUsbDevice, OpenBsdGlobalMemory (abstract base),
  OpenBsdVirtualMemory

Remaining in oshi-core (have native calls):
- OpenBsdSysctlUtil, OpenBsdCentralProcessor, OpenBsdGlobalMemory,
  OpenBsdHWDiskStore, OpenBsdOSProcess, OpenBsdOperatingSystem,
  OpenBsdFileSystem, OpenBsdComputerSystem, OpenBsdPowerSource,
  OpenBsdSensors
