+++ oshi-common/src/main/java/module-info.java
@@ -32,17 +32,20 @@ module com.github.oshi.common {
     exports oshi.hardware.common.platform.mac;
     exports oshi.hardware.common.platform.unix;
     exports oshi.hardware.common.platform.unix.freebsd;
+    exports oshi.hardware.common.platform.unix.openbsd;
     exports oshi.hardware.common.platform.windows;
     exports oshi.nativefree;
     exports oshi.software.common;
     exports oshi.software.common.os.linux;
     exports oshi.software.common.os.mac;
     exports oshi.software.common.os.unix.freebsd;
+    exports oshi.software.common.os.unix.openbsd;
     exports oshi.software.common.os.windows;
     exports oshi.software.os;
     exports oshi.spi;
     exports oshi.util;
     exports oshi.util.common.platform.unix.freebsd;
+    exports oshi.util.common.platform.unix.openbsd;
     exports oshi.util.driver.linux;
     exports oshi.util.driver.linux.proc;
     exports oshi.util.driver.unix;
