+++ oshi-common/src/main/java/oshi/hardware/common/platform/unix/openbsd/OpenBsdGraphicsCard.java
@@ -2,7 +2,7 @@
  * Copyright 2021-2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
-package oshi.hardware.platform.unix.openbsd;
+package oshi.hardware.common.platform.unix.openbsd;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -17,24 +17,15 @@ import oshi.util.Constants;
 import oshi.util.ExecutingCommand;
 
 /**
- * Graphics Card info obtained from pciconf
+ * Graphics Card info obtained from pcidump
  */
 @Immutable
-final class OpenBsdGraphicsCard extends AbstractGraphicsCard {
+public class OpenBsdGraphicsCard extends AbstractGraphicsCard {
 
     private static final String PCI_CLASS_DISPLAY = "Class: 03 Display";
     private static final Pattern PCI_DUMP_HEADER = Pattern.compile(" \\d+:\\d+:\\d+: (.+)");
 
-    /**
-     * Constructor for OpenBsdGraphicsCard
-     *
-     * @param name        The name
-     * @param deviceId    The device ID
-     * @param vendor      The vendor
-     * @param versionInfo The version info
-     * @param vram        The VRAM
-     */
-    OpenBsdGraphicsCard(String name, String deviceId, String vendor, String versionInfo, long vram) {
+    public OpenBsdGraphicsCard(String name, String deviceId, String vendor, String versionInfo, long vram) {
         super(name, deviceId, vendor, versionInfo, vram);
     }
 
@@ -45,7 +36,6 @@ final class OpenBsdGraphicsCard extends AbstractGraphicsCard {
      */
     public static List<GraphicsCard> getGraphicsCards() {
         List<GraphicsCard> cardList = new ArrayList<>();
-        // Enumerate all devices and add if required
         List<String> devices = ExecutingCommand.runNative("pcidump -v");
         if (devices.isEmpty()) {
             return Collections.emptyList();
@@ -58,25 +48,18 @@ final class OpenBsdGraphicsCard extends AbstractGraphicsCard {
         for (String line : devices) {
             Matcher m = PCI_DUMP_HEADER.matcher(line);
             if (m.matches()) {
-                // Identifies start of a new device. Save previous if it's a graphics card
                 if (classCodeFound) {
                     cardList.add(new OpenBsdGraphicsCard(name.isEmpty() ? Constants.UNKNOWN : name,
                             productId.isEmpty() ? "0x0000" : productId, vendorId.isEmpty() ? "0x0000" : vendorId,
                             versionInfo.isEmpty() ? Constants.UNKNOWN : versionInfo, 0L));
                 }
-                // Device name is the captured pattern
                 name = m.group(1);
-                // Reset values
                 vendorId = "";
                 productId = "";
                 classCodeFound = false;
                 versionInfo = "";
             } else {
                 int idx;
-                // Look for:
-                // 0x0000: Vendor ID: 1ab8, Product ID: 4005
-                // 0x0008: Class: 03 Display, Subclass: 00 VGA
-                // ....... Interface: 00, Revision: 00
                 if (!classCodeFound) {
                     idx = line.indexOf("Vendor ID: ");
                     if (idx >= 0 && line.length() >= idx + 15) {
@@ -97,7 +80,6 @@ final class OpenBsdGraphicsCard extends AbstractGraphicsCard {
                 }
             }
         }
-        // In case we reached end before saving
         if (classCodeFound) {
             cardList.add(new OpenBsdGraphicsCard(name.isEmpty() ? Constants.UNKNOWN : name,
                     productId.isEmpty() ? "0x0000" : productId, vendorId.isEmpty() ? "0x0000" : vendorId,
