+++ oshi-common/src/main/java/oshi/hardware/common/platform/unix/openbsd/OpenBsdUsbDevice.java
@@ -2,7 +2,7 @@
  * Copyright 2021-2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
-package oshi.hardware.platform.unix.openbsd;
+package oshi.hardware.common.platform.unix.openbsd;
 
 import static java.util.Collections.sort;
 
@@ -30,10 +30,6 @@ public class OpenBsdUsbDevice extends AbstractUsbDevice {
     /**
      * Instantiates a list of {@link oshi.hardware.UsbDevice} objects, representing devices connected via a usb port
      * (including internal devices).
-     * <p>
-     * If the value of {@code tree} is true, the top level devices returned from this method are the USB Controllers;
-     * connected hubs and devices in its device tree share that controller's bandwidth. If the value of {@code tree} is
-     * false, USB devices (not controllers) are listed in a single flat list.
      *
      * @param tree If true, returns a list of controllers, which requires recursive iteration of connected devices. If
      *             false, returns a flat list of devices excluding controllers.
@@ -52,8 +48,6 @@ public class OpenBsdUsbDevice extends AbstractUsbDevice {
     }
 
     private static List<UsbDevice> getUsbDevices() {
-        // Maps to store information using node # as the key
-        // Node is controller+addr (+port+addr etc.)
         Map<String, String> nameMap = new HashMap<>();
         Map<String, String> vendorMap = new HashMap<>();
         Map<String, String> vendorIdMap = new HashMap<>();
@@ -62,23 +56,16 @@ public class OpenBsdUsbDevice extends AbstractUsbDevice {
         Map<String, List<String>> hubMap = new HashMap<>();
 
         List<String> rootHubs = new ArrayList<>();
-        // For each item enumerated, store information in the maps
         String key = "";
-        // Addresses repeat for each controller:
-        // prepend the controller /dev/usb* for the key
         String parent = "";
-        // Enumerate all devices and build information maps.
-        // This will build the entire device tree in hubMap
         for (String line : ExecutingCommand.runNative("usbdevs -v")) {
             if (line.startsWith("Controller ")) {
                 parent = line.substring(11);
             } else if (line.startsWith("addr ")) {
-                // addr 01: 8086:0000 Intel, EHCI root hub
                 if (line.indexOf(':') == 7 && line.indexOf(',') >= 18) {
                     key = parent + line.substring(0, 7);
                     String[] split = line.substring(8).trim().split(",");
                     if (split.length > 1) {
-                        // 0 = vid:pid vendor
                         String vendorStr = split[0].trim();
                         int idx1 = vendorStr.indexOf(':');
                         int idx2 = vendorStr.indexOf(' ');
@@ -87,11 +74,8 @@ public class OpenBsdUsbDevice extends AbstractUsbDevice {
                             productIdMap.put(key, vendorStr.substring(idx1 + 1, idx2));
                             vendorMap.put(key, vendorStr.substring(idx2 + 1));
                         }
-                        // 1 = product
                         nameMap.put(key, split[1].trim());
-                        // Add this key to the parent's hubmap list
                         hubMap.computeIfAbsent(parent, x -> new ArrayList<>()).add(key);
-                        // For the first addr in a controller, make it the parent
                         if (!parent.contains("addr")) {
                             parent = key;
                             rootHubs.add(parent);
@@ -99,9 +83,6 @@ public class OpenBsdUsbDevice extends AbstractUsbDevice {
                     }
                 }
             } else if (!key.isEmpty()) {
-                // Continuing to read for the previous key
-                // CSV is speed, power, config, rev, optional iSerial
-                // Since all we need is the serial...
                 int idx = line.indexOf("iSerial ");
                 if (idx >= 0) {
                     serialMap.put(key, line.substring(idx + 8).trim());
@@ -110,7 +91,6 @@ public class OpenBsdUsbDevice extends AbstractUsbDevice {
             }
         }
 
-        // Build tree and return
         List<UsbDevice> controllerDevices = new ArrayList<>();
         for (String devusb : rootHubs) {
             controllerDevices.add(getDeviceAndChildren(devusb, "0000", "0000", nameMap, vendorMap, vendorIdMap,
@@ -119,20 +99,6 @@ public class OpenBsdUsbDevice extends AbstractUsbDevice {
         return controllerDevices;
     }
 
-    /**
-     * Recursively creates OpenBsdUsbDevices by fetching information from maps to populate fields
-     *
-     * @param devPath      The device node path.
-     * @param vid          The default (parent) vendor ID
-     * @param pid          The default (parent) product ID
-     * @param nameMap      the map of names
-     * @param vendorMap    the map of vendors
-     * @param vendorIdMap  the map of vendorIds
-     * @param productIdMap the map of productIds
-     * @param serialMap    the map of serial numbers
-     * @param hubMap       the map of hubs
-     * @return An OpenBsdUsbDevice corresponding to this device
-     */
     private static OpenBsdUsbDevice getDeviceAndChildren(String devPath, String vid, String pid,
             Map<String, String> nameMap, Map<String, String> vendorMap, Map<String, String> vendorIdMap,
             Map<String, String> productIdMap, Map<String, String> serialMap, Map<String, List<String>> hubMap) {
