+++ oshi-common/src/main/java/oshi/hardware/common/platform/unix/openbsd/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides abstract base classes and pure-Java implementations for OpenBSD hardware, shared by the JNA and FFM modules.
+ */
+package oshi.hardware.common.platform.unix.openbsd;
