+++ oshi-common/src/main/java/oshi/software/common/os/unix/openbsd/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Provides abstract base classes and pure-Java implementations for the OpenBSD operating system, shared by the JNA and
+ * FFM modules.
+ */
+package oshi.software.common.os.unix.openbsd;
