+++ oshi-common/src/main/java/oshi/util/common/platform/unix/openbsd/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides pure-Java utility classes for OpenBSD, shared by the JNA and FFM modules.
+ */
+package oshi.util.common.platform.unix.openbsd;
