+++ oshi-common/src/test/java/module-info.test
@@ -52,6 +52,12 @@
   com.github.oshi.common/oshi.hardware.common.platform.unix.freebsd=org.junit.platform.commons
 --add-opens
   com.github.oshi.common/oshi.util.common.platform.unix.freebsd=org.junit.platform.commons
+--add-opens
+  com.github.oshi.common/oshi.hardware.common.platform.unix.openbsd=org.junit.platform.commons
+--add-opens
+  com.github.oshi.common/oshi.software.common.os.unix.openbsd=org.junit.platform.commons
+--add-opens
+  com.github.oshi.common/oshi.util.common.platform.unix.openbsd=org.junit.platform.commons
 
 --add-modules
   org.hamcrest,org.slf4j.simple
